--- src/GraphQL/query/QueryGetWork.ts
@@ -1,15 +1,0 @@
-import gql from "graphql-tag";
-
-export default gql(`
-    query　{
-        gethWorkByTag(tags:"tag") {
-            items{
-                id
-                tags
-                userId
-                title
-            },
-            nextToken
-        }
-    }
-`);
