+++ src/components/ComposingRoute.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React                 from "react";
 import { Route, RouteProps } from "react-router-dom";
 
 export default (
