+++ src/components/ComposingSwitch.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode } from "react";
-import { Switch } from "react-router-dom";
+import { Switch }           from "react-router-dom";
 
 interface Props {
     children: ReactNode;
