+++ src/components/GraphQLProgress.tsx
@@ -1,6 +1,6 @@
-import React from "react";
-import styled from "styled-components";
+import React                                       from "react";
 import CircularProgress, { CircularProgressProps } from "@material-ui/core/CircularProgress";
+import styled                                      from "styled-components";
 
 export default (props: CircularProgressProps) => (
     <Host>
