+++ src/components/ImageInput.tsx
@@ -1,6 +1,6 @@
 import React, { Fragment } from "react";
-import Image from "./Image";
-import styled from "styled-components";
+import styled              from "styled-components";
+import Image               from "./Image";
 
 interface PropsBase {
     defaultImageUrl?: string;
