+++ src/components/wrapper/AppSyncClient.tsx
@@ -1,12 +1,12 @@
 declare const require: any;
 import React, { ReactChild } from "react";
-// https://github.com/awslabs/aws-mobile-appsync-sdk-js/pull/141
+// https://github.com/awslabs/aws-mobile-appsync-sdk-js/pull/155
 // https://github.com/awslabs/aws-mobile-appsync-sdk-js/issues/48
 const { Rehydrated }       = require("aws-appsync-react");
 const { AUTH_TYPE }        = require("aws-appsync/lib/link/auth-link");
 const { AWSAppSyncClient } = require("aws-appsync");
 import { ApolloProvider }   from "react-apollo";
-import config from "../../config";
+import config        from "../../config";
 import { AuthProps } from "./Auth";
 
 interface Props extends AuthProps {
