+++ src/components/wrapper/NotificationListener.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode, Fragment } from "react";
-import NotificationComponent from "../Notification";
+import NotificationComponent          from "../Notification";
 
 export type Notification = (type: "info" | "error", message: string) => void;
 export type ErrorNotification = (error: Error) => void;
